+++ packages/api-headless-cms/src/content/apolloHandler/createApolloHandler.ts
@@ -1,17 +1,8 @@
-import { ApolloServer } from "apollo-server-lambda";
-import { boolean } from "boolean";
 import { CreateSchemaPlugin } from "@webiny/handler-apollo-server/types";
-import { get } from "lodash";
 import { HandlerContext } from "@webiny/handler/types";
-
-function normalizeEvent(event) {
-    // In AWS, when enabling binary support, received body gets base64 encoded. Did not find a way to solve this
-    // correctly, so for now we "normalize" the event before passing it to the handler. It would be nice if
-    // we could resolve this issue better / smarter in the future (configure integrations correctly?).
-    if (event.isBase64Encoded) {
-        event.body = Buffer.from(event.body, "base64").toString("utf-8");
-    }
-}
+import { generateSchemaHash } from "apollo-server-core/dist/utils/schemaHash";
+import { runHttpQuery as apolloRunHttpQuery } from "apollo-server-core/dist/runHttpQuery";
+import { Headers } from "apollo-server-env";
 
 type CreateApolloHandlerParams = {
     context: HandlerContext;
@@ -21,11 +12,8 @@ type CreateApolloHandlerParams = {
 
 export default async function createApolloHandler({
     context,
-    options,
     meta = null
 }: CreateApolloHandlerParams) {
-    const { server = {}, handler = {} } = options;
-
     const createSchemaPlugin = context.plugins.byName<CreateSchemaPlugin>(
         "handler-apollo-server-create-schema"
     );
@@ -35,63 +23,41 @@ export default async function createApolloHandler({
     }
 
     const { schema } = await createSchemaPlugin.create(context);
+    const schemaHash = generateSchemaHash(schema);
 
-    const apolloServer = new ApolloServer({
-        // @ts-ignore Not sure why it doesn't work, "boolean" function does return a boolean value.
-        introspection: boolean(server.introspection),
-        // @ts-ignore Not sure why it doesn't work, "boolean" function does return a boolean value.
-        playground: boolean(server.playground),
-        debug: boolean(process.env.DEBUG),
-        ...server,
-        schema,
-        context: ({ context }) => context
-    });
-
-    const apolloHandler = apolloServer.createHandler({
-        cors: {
-            origin: "*",
-            methods: "GET,HEAD,POST",
-            ...(handler.cors || {})
-        }
-    });
+    const runHttpQuery = (query, context) => {
+        return apolloRunHttpQuery([], {
+            method: "POST",
+            query,
+            options: { schema, schemaHash, context },
+            request: {
+                url: context.http.path.base,
+                method: context.http.method,
+                headers: new Headers(context.http.headers)
+            }
+        });
+    };
 
     const output = {
         meta,
+        schema,
+        handler: runHttpQuery,
         async queryMeta() {
             // Invoke handler and check if environment / environment alias has changed.
-            const response = await output.handler(
+            const { graphqlResponse } = await runHttpQuery(
                 {
-                    isMetaQuery: true,
-                    httpMethod: "POST",
-                    headers: {},
-                    body: JSON.stringify({
-                        query: /* GraphQL */ `
-                            {
-                                getMeta {
-                                    cacheKey
-                                }
+                    query: /* GraphQL */ `
+                        {
+                            getMeta {
+                                cacheKey
                             }
-                        `
-                    })
+                        }
+                    `
                 },
                 context
             );
 
-            const body = JSON.parse(response.body);
-            return get(body, "data.getMeta") || {};
-        },
-        schema,
-        handler: (event, context): any => {
-            normalizeEvent(event);
-            return new Promise((resolve, reject) => {
-                apolloHandler(event, context, (error, data) => {
-                    if (error) {
-                        return reject(error);
-                    }
-
-                    resolve(data);
-                });
-            });
+            return JSON.parse(graphqlResponse)?.data?.getMeta;
         }
     };