+++ packages/cli/cli.js
@@ -1,17 +1,6 @@
 #!/usr/bin/env node
 const yargs = require("yargs");
 const { blue } = require("chalk");
-const indentString = require("indent-string");
-
-// Add indentation to console.log output
-const log = console.log;
-console.log = (first = "", ...args) => {
-    if (typeof first === "string") {
-        log(indentString(first, 2), ...args);
-    } else {
-        log(first, ...args);
-    }
-};
 
 const { createCommands } = require("./commands");
 const context = require("./context");