+++ packages/handler-args/src/types.ts
@@ -1,0 +1,3 @@
+export type HandlerArgsContext = {
+    invocationArgs: { [key: string]: any };
+};