+++ packages/handler/src/index.ts
@@ -1,2 +1,1 @@
-export { default as createHandler } from "./create";
-export { default as createResponse } from "./createResponse";
+export { default as createHandler } from "./createHandler";