+++ packages/api-i18n/src/locales/index.ts
@@ -18,6 +18,5 @@ export default () => [
                 createdOn: locale.createdOn
             }));
         }
-    } as HandlerPlugin,
-
+    } as HandlerPlugin
 ];